Safari-611.1.10/LayoutTests/accessibility/mac/selection-sync-expected.txt

one two three. four five six.

This tests that selection changes originating from accessibility API include AXTextStateSync=true.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS addedNotification is true
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeSelectionMove
PASS results[resultIndex]["AXTextSelectionDirection"] is AXTextSelectionDirectionDiscontiguous
PASS results[resultIndex]["AXTextStateSync"] is true
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeSelectionMove
PASS results[resultIndex]["AXTextSelectionDirection"] is AXTextSelectionDirectionDiscontiguous
PASS results[resultIndex]["AXTextStateSync"] is true
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeSelectionExtend
PASS results[resultIndex]["AXTextSelectionDirection"] is AXTextSelectionDirectionDiscontiguous
PASS results[resultIndex]["AXTextStateSync"] is true
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeSelectionMove
PASS results[resultIndex]["AXTextSelectionDirection"] is AXTextSelectionDirectionDiscontiguous
PASS results[resultIndex]["AXTextStateSync"] is true
PASS results[resultIndex]["AXTextStateChangeType"] is AXTextStateChangeTypeSelectionExtend
PASS results[resultIndex]["AXTextSelectionDirection"] is AXTextSelectionDirectionDiscontiguous
PASS results[resultIndex]["AXTextStateSync"] is true
PASS successfullyParsed is true

TEST COMPLETE
